Consolidate settings form state and save coordination
