+++ app/static/js/settings.js
@@ -1,2231 +1,35 @@
-/* ═══════════════════════════════════════════════
-   DOCSight – Settings JavaScript
-   Settings behavior + Mobile List→Detail
-   ═══════════════════════════════════════════════ */
-
-/* Globals validated and installed by settings-bootstrap.js: T, SECTION_TITLES,
-   serverOffsetMin, serverTz, currentLang, currentTz, savedCooldowns */
-
-/* ── Section Controller ── */
-var _currentSection = 'connection';
-var _saveHiddenSections = { support: true, about: true };
-var _mobileSidebarFocusReturn = null;
-var _mobileSidebarMedia = window.matchMedia ? window.matchMedia('(max-width: 768px)') : null;
-
-function _isMobileSidebarMode() {
-    return !!(_mobileSidebarMedia && _mobileSidebarMedia.matches);
-}
-
-function _setSidebarAccessibility(sidebar, isOpen) {
-    if (!sidebar) return;
-    if (_isMobileSidebarMode() && !isOpen) {
-        sidebar.setAttribute('aria-hidden', 'true');
-        sidebar.setAttribute('inert', '');
-    } else {
-        sidebar.removeAttribute('aria-hidden');
-        sidebar.removeAttribute('inert');
+/* Settings composition and the explicit boundary for template/module handlers. */
+(function(S) {
+    'use strict';
+    var navigation;
+    var form = S.form({state: S.state, showsSaveFooter: function() { return navigation.showsSaveFooter(); }});
+    var tokens = S.tokens({showToast: form.showToast});
+    var connections = S.connections({getFormData: form.getFormData});
+    var backups = S.backups({showToast: form.showToast});
+    var themes = S.themes({showToast: form.showToast, guardUnsaved: form.guardUnsaved});
+    var smartCapture = S.smart_capture({showToast: form.showToast});
+    var registry = S.module_registry({showToast: form.showToast});
+    navigation = S.navigation({syncSaveFooter: form.syncSaveFooter, onSection: function(id, panel) {
+        if (id === 'security') tokens.loadApiTokens();
+        if (panel && panel.querySelector('#backup-list')) backups.loadBackupList();
+        if (id === 'appearance') themes.loadThemeRegistryIfNeeded();
+        if (id === 'smart_capture') smartCapture.loadSmartCaptureHistory();
+        if (id === 'extensions') registry.refreshModuleRegistry();
+    }});
+    var notifications = S.notifications({showToast: form.showToast, saveInstantly: form.saveInstantly, syncCard: navigation.syncCard});
+    function expose(owner, names) {
+        names.split(' ').forEach(function(name) { window[name] = owner[name]; });
     }
-}
-
-function syncMobileSidebarAccessibility() {
-    var sidebar = document.getElementById('settings-sidebar');
-    _setSidebarAccessibility(sidebar, !!(sidebar && sidebar.classList.contains('open')));
-}
-
-function _applySection(id) {
-    _currentSection = id;
-
-    /* Sidebar: update active link */
-    document.querySelectorAll('.nav-item[data-section]').forEach(function(link) {
-        var isActive = link.getAttribute('data-section') === id;
-        link.classList.toggle('active', isActive);
-        if (isActive) {
-            link.setAttribute('aria-current', 'page');
-        } else {
-            link.removeAttribute('aria-current');
-        }
-    });
-
-    /* Panels: show selected */
-    document.querySelectorAll('.settings-panel').forEach(function(panel) {
-        panel.classList.remove('active');
-    });
-    var target = document.getElementById('panel-' + id);
-    if (target) target.classList.add('active');
-
-    /* Mobile title */
-    var mobileTitle = document.getElementById('mobile-title');
-    if (mobileTitle) mobileTitle.textContent = SECTION_TITLES[id] || id;
-
-    /* Save footer: hide on support/modules, otherwise respect dirty state */
-    _syncSaveFooter();
-
-    /* Auto-load data for certain panels */
-    if (id === 'security') loadApiTokens();
-    if (target && target.querySelector('#backup-list')) loadBackupList();
-    if (id === 'appearance') loadThemeRegistryIfNeeded();
-    if (id === 'smart_capture') loadSmartCaptureHistory();
-    if (id === 'extensions') refreshModuleRegistry();
-
-    /* Mobile: close sidebar after selection */
-    closeMobileSidebar();
-}
-
-/* User-initiated section change: apply and add a browser history entry so
-   Back/Forward move between previously viewed settings sections. Re-selecting
-   the current section replaces state instead of stacking a duplicate entry. */
-function switchSection(id) {
-    var isNewSection = id !== _currentSection;
-    _applySection(id);
-    if (isNewSection) {
-        history.pushState(null, '', '#' + id);
-    } else {
-        history.replaceState(null, '', '#' + id);
-    }
-}
-
-/* Resolve the section referenced by the current URL hash (default: connection). */
-function _sectionFromHash() {
-    var hash = location.hash.replace('#', '');
-    return (hash && document.getElementById('panel-' + hash)) ? hash : 'connection';
-}
-
-/* Non-pushing variant for Back/Forward and manual hash edits: the browser has
-   already updated the URL, so only reflect it in the UI. The guard also makes
-   the popstate+hashchange double-fire on navigation a no-op the second time. */
-function _syncSectionFromHash() {
-    var id = _sectionFromHash();
-    if (id === _currentSection) return;
-    _applySection(id);
-}
-
-window.addEventListener('popstate', _syncSectionFromHash);
-window.addEventListener('hashchange', _syncSectionFromHash);
-
-/* ── Mobile Sidebar ── */
-function openMobileSidebar() {
-    var sidebar = document.getElementById('settings-sidebar');
-    var backdrop = document.getElementById('sidebar-backdrop');
-    var menuButton = document.getElementById('mobile-menu-button');
-    _mobileSidebarFocusReturn = document.activeElement || menuButton;
-    if (sidebar) {
-        sidebar.classList.add('open');
-        _setSidebarAccessibility(sidebar, true);
-    }
-    if (backdrop) backdrop.classList.add('active');
-    if (menuButton) menuButton.setAttribute('aria-expanded', 'true');
-    var activeNav = sidebar ? sidebar.querySelector('.nav-item.active[data-section]') : null;
-    var firstNav = sidebar ? sidebar.querySelector('.nav-item[data-section]') : null;
-    var focusTarget = activeNav || firstNav;
-    if (focusTarget && typeof focusTarget.focus === 'function') focusTarget.focus();
-}
-
-function closeMobileSidebar(options) {
-    options = options || {};
-    var sidebar = document.getElementById('settings-sidebar');
-    var backdrop = document.getElementById('sidebar-backdrop');
-    var menuButton = document.getElementById('mobile-menu-button');
-    var wasOpen = !!(sidebar && sidebar.classList.contains('open'));
-    if (sidebar) {
-        sidebar.classList.remove('open');
-        _setSidebarAccessibility(sidebar, false);
-    }
-    if (backdrop) backdrop.classList.remove('active');
-    if (menuButton) menuButton.setAttribute('aria-expanded', 'false');
-    var shouldRestoreFocus = options.restoreFocus !== false;
-    if (wasOpen && shouldRestoreFocus) {
-        var focusTarget = _mobileSidebarFocusReturn;
-        if (!focusTarget || typeof focusTarget.focus !== 'function' || !document.contains(focusTarget)) {
-            focusTarget = menuButton;
-        }
-        if (focusTarget && typeof focusTarget.focus === 'function') focusTarget.focus();
-    }
-    _mobileSidebarFocusReturn = null;
-}
-
-document.addEventListener('keydown', function(event) {
-    if (event.key !== 'Escape') return;
-    var sidebar = document.getElementById('settings-sidebar');
-    if (sidebar && sidebar.classList.contains('open')) {
-        event.preventDefault();
-        closeMobileSidebar();
-    }
-});
-
-/* ── Collapsible Cards ── */
-function _syncCardCollapseAria(card) {
-    if (!card) return;
-    var expanded = !card.classList.contains('collapsed');
-    card.querySelectorAll('.card-collapse-toggle[aria-expanded]').forEach(function(toggle) {
-        toggle.setAttribute('aria-expanded', expanded ? 'true' : 'false');
-    });
-    if (card.classList.contains('notification-channel-card')) {
-        var body = card.querySelector('.card-collapse-body');
-        if (body) {
-            body.setAttribute('aria-hidden', expanded ? 'false' : 'true');
-            if (expanded) {
-                body.removeAttribute('inert');
-            } else {
-                body.setAttribute('inert', '');
-            }
-        }
-    }
-}
-
-function toggleCardCollapse(headerEl) {
-    var card = headerEl.closest('.collapsible-card');
-    if (card) {
-        card.classList.toggle('collapsed');
-        _syncCardCollapseAria(card);
+    expose(form, 'getFormData showToast');
+    expose(navigation, 'switchSection openMobileSidebar closeMobileSidebar toggleCardCollapse');
+    expose(tokens, 'createApiToken copyToken');
+    expose(connections, 'onIspChange testModem testMqtt testSpeedtest clearSpeedtestCache testNotifications');
+    expose(backups, 'downloadBackup backupNow openBrowseModal closeBrowseModal selectBrowsePath');
+    expose(themes, 'toggleThemeFromAppearance applyFontToggle previewTheme cancelPreview applyPreviewedTheme applyTheme refreshRegistry');
+    expose(notifications, 'subscribePwaPush unsubscribePwaPush');
+    expose(registry, 'refreshModuleRegistry');
+    document.addEventListener('DOMContentLoaded', function() {
         if (typeof lucide !== 'undefined') lucide.createIcons();
-    }
-}
-
-function expandNotificationChannelCard(controlEl) {
-    if (!controlEl || !controlEl.checked) return;
-    var card = controlEl.closest('.notification-channel-card');
-    if (card && card.classList.contains('collapsed')) {
-        card.classList.remove('collapsed');
-        _syncCardCollapseAria(card);
-    }
-}
-
-function _notificationLabel(key, fallback) {
-    return (typeof T !== 'undefined' && T[key]) ? T[key] : fallback;
-}
-
-function _setNotificationChannelBadge(channel, enabled, label) {
-    var badge = document.querySelector('[data-channel-badge="' + channel + '"]');
-    if (!badge) return;
-    badge.textContent = label;
-    badge.classList.toggle('status-enabled', enabled);
-    badge.classList.toggle('status-disabled', !enabled && channel !== 'webhook');
-    badge.classList.toggle('status-muted', !enabled && channel === 'webhook');
-}
-
-function updateNotificationChannelSummaries() {
-    var webhook = document.getElementById('notify_webhook_url');
-    var webhookConfigured = !!(webhook && webhook.value.trim() !== '');
-    _setNotificationChannelBadge(
-        'webhook',
-        webhookConfigured,
-        webhookConfigured ? _notificationLabel('enabled', 'Enabled') : _notificationLabel('not_configured', 'Not configured')
-    );
-
-    var apprise = document.getElementById('notify_apprise_enabled');
-    var appriseEnabled = !!(apprise && apprise.checked);
-    _setNotificationChannelBadge(
-        'apprise',
-        appriseEnabled,
-        appriseEnabled ? _notificationLabel('enabled', 'Enabled') : _notificationLabel('disabled', 'Disabled')
-    );
-
-    var pwa = document.getElementById('notify_pwa_push_enabled');
-    var pwaEnabled = !!(pwa && pwa.checked);
-    _setNotificationChannelBadge(
-        'pwa',
-        pwaEnabled,
-        pwaEnabled ? _notificationLabel('enabled', 'Enabled') : _notificationLabel('disabled', 'Disabled')
-    );
-}
-
-/* ── API Token Management ── */
-function _tokenCell(text, style) {
-    var td = document.createElement('td');
-    td.style.cssText = style || 'padding:4px 8px;';
-    td.textContent = text;
-    return td;
-}
-
-function loadApiTokens() {
-    var table = document.getElementById('api-tokens-table');
-    var body = document.getElementById('api-tokens-body');
-    var empty = document.getElementById('api-tokens-empty');
-    if (!table || !body) return;
-    fetch(docsightUrl('/api/tokens')).then(function(r) { return r.json(); }).then(function(data) {
-        var tokens = (data.tokens || []).filter(function(t) { return !t.revoked; });
-        while (body.firstChild) body.removeChild(body.firstChild);
-        if (tokens.length === 0) {
-            table.style.display = 'none';
-            if (empty) empty.style.display = 'block';
-            return;
-        }
-        table.style.display = 'table';
-        if (empty) empty.style.display = 'none';
-        tokens.forEach(function(tk) {
-            var tr = document.createElement('tr');
-            tr.appendChild(_tokenCell(tk.name, 'padding:4px 8px;'));
-            var prefixTd = document.createElement('td');
-            prefixTd.style.cssText = 'padding:4px 8px;';
-            var code = document.createElement('code');
-            code.textContent = tk.token_prefix + '...';
-            prefixTd.appendChild(code);
-            tr.appendChild(prefixTd);
-            tr.appendChild(_tokenCell(tk.last_used_at || '\u2014', 'padding:4px 8px;'));
-            var actionTd = document.createElement('td');
-            actionTd.style.cssText = 'padding:4px 8px;';
-            var btn = document.createElement('button');
-            btn.type = 'button';
-            btn.className = 'btn btn-sm';
-            btn.style.cssText = 'font-size:0.8em;padding:2px 8px;';
-            btn.textContent = T.api_token_revoke || 'Revoke';
-            btn.setAttribute('data-token-id', tk.id);
-            btn.setAttribute('data-token-name', tk.name);
-            btn.addEventListener('click', function() {
-                revokeToken(parseInt(this.getAttribute('data-token-id')), this.getAttribute('data-token-name'));
-            });
-            actionTd.appendChild(btn);
-            tr.appendChild(actionTd);
-            body.appendChild(tr);
-        });
-    }).catch(function() {});
-}
-
-function createApiToken() {
-    var inp = document.getElementById('api-token-name');
-    var name = (inp.value || '').trim();
-    if (!name) { inp.focus(); return; }
-    fetch(docsightUrl('/api/tokens'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify({name: name})
-    }).then(function(r) { return r.json().then(function(d) { return {ok: r.ok, data: d}; }); })
-    .then(function(res) {
-        if (!res.ok) { showToast(res.data.error || (T.error_prefix || 'Error'), false); return; }
-        inp.value = '';
-        var banner = document.getElementById('api-token-created-banner');
-        document.getElementById('api-token-plaintext').textContent = res.data.token;
-        banner.style.display = 'block';
-        loadApiTokens();
-        if (typeof lucide !== 'undefined') lucide.createIcons();
-    }).catch(function() { showToast(T.error_prefix || 'Error', false); });
-}
-
-function copyToken() {
-    var text = document.getElementById('api-token-plaintext').textContent;
-    navigator.clipboard.writeText(text).then(function() {
-        showToast(T.api_token_copied || 'Token copied!', true);
+        [themes, connections, backups, notifications, smartCapture, form, navigation].forEach(function(owner) { owner.init(); });
     });
-}
-
-function revokeToken(id, name) {
-    var msg = (T.api_token_revoke_confirm || 'Revoke token "{name}"?').replace('{name}', name);
-    docsightConfirm({
-        title: T.api_token_revoke || 'Revoke token',
-        message: msg,
-        confirmText: T.revoke || 'Revoke',
-        cancelText: T.cancel || 'Cancel',
-        danger: true
-    }).then(function(confirmed) {
-        if (!confirmed) return null;
-        return fetch(docsightUrl('/api/tokens/' + id), {method: 'DELETE'});
-    })
-    .then(function(r) { return r ? r.json() : null; })
-    .then(function(data) {
-        if (!data) return;
-        if (data.success) {
-            showToast(T.api_token_revoked || 'Token revoked', true);
-            document.getElementById('api-token-created-banner').style.display = 'none';
-            loadApiTokens();
-        } else {
-            showToast(data.error || (T.error_prefix || 'Error'), false);
-        }
-    }).catch(function() { showToast(T.error_prefix || 'Error', false); });
-}
-
-/* ── Connection Test Status Indicators ── */
-function setConnectionTestStatus(boxId, labelId, state, text) {
-    var box = document.getElementById(boxId);
-    if (!box) return;
-    box.hidden = false;
-    box.classList.remove('testing', 'connected', 'disconnected');
-    box.classList.add(state);
-    var label = document.getElementById(labelId);
-    if (label) label.textContent = text;
-}
-
-function setModemStatus(state, text) {
-    setConnectionTestStatus('modem-status', 'modem-status-text', state, text);
-}
-
-function setMqttStatus(state, text) {
-    setConnectionTestStatus('mqtt-status', 'mqtt-status-text', state, text);
-}
-
-/* ── Theme Toggle ── */
-function initThemeToggle() {
-    var appearanceCheck = document.getElementById('theme-toggle-appearance');
-
-    /* Restore saved theme */
-    var saved = localStorage.getItem('docsis-theme');
-    if (saved) {
-        document.documentElement.setAttribute('data-theme', saved);
-        if (appearanceCheck) appearanceCheck.checked = (saved === 'dark');
-    }
-}
-
-function toggleThemeFromAppearance(checked) {
-    var theme = checked ? 'dark' : 'light';
-    document.documentElement.setAttribute('data-theme', theme);
-    localStorage.setItem('docsis-theme', theme);
-    updatePaletteDots(theme);
-}
-
-function applyFontToggle(useSystem) {
-    var el = document.getElementById('font-override');
-    var hidden = document.getElementById('font_family');
-    if (hidden) hidden.value = useSystem ? 'system' : 'outfit';
-    if (useSystem) {
-        if (!el) {
-            el = document.createElement('style');
-            el.id = 'font-override';
-            el.textContent = ':root { --font-sans: system-ui, -apple-system, "Segoe UI", Roboto, sans-serif; }';
-            document.head.appendChild(el);
-        }
-    } else if (el) {
-        el.remove();
-    }
-}
-
-function updatePaletteDots(mode) {
-    var keys = ['--bg', '--surface', '--accent', '--good', '--crit'];
-    document.querySelectorAll('.theme-card').forEach(function(card) {
-        var colors = JSON.parse(card.getAttribute('data-theme-' + mode) || '{}');
-        var dots = card.querySelectorAll('.palette-dot');
-        dots.forEach(function(dot, i) {
-            if (keys[i] && colors[keys[i]]) dot.style.background = colors[keys[i]];
-        });
-    });
-}
-
-/* ── ISP Change ── */
-function onIspChange() {
-    var sel = document.getElementById('isp_select');
-    var row = document.getElementById('isp-other-row');
-    var icon = document.getElementById('isp-icon-preview');
-    if (!sel) return;
-    row.style.display = sel.value === '__other__' ? 'flex' : 'none';
-    var isp = sel.value.toLowerCase();
-    var iconMap = {
-        'vodafone': '/static/img/providers/vodafone.svg',
-        'telekom': '/static/img/providers/telekom.svg',
-        'o2': '/static/img/providers/o2.svg'
-    };
-    if (sel.value && sel.value !== '__other__') {
-        icon.src = docsightUrl(iconMap[isp] || '/static/img/providers/generic.svg');
-        icon.alt = sel.value;
-        icon.style.display = 'block';
-        icon.style.opacity = iconMap[isp] ? '1' : '0.7';
-    } else {
-        icon.style.display = 'none';
-    }
-}
-
-/* ── Toast ── */
-function showToast(msg, ok) {
-    var el = document.getElementById('toast');
-    el.textContent = msg;
-    el.className = 'toast ' + (ok ? 'toast-ok' : 'toast-fail');
-    el.style.display = 'block';
-    setTimeout(function() { el.style.display = 'none'; }, 3000);
-}
-
-/* ── Form Data ── */
-var MASK = '\u2022\u2022\u2022\u2022\u2022\u2022\u2022\u2022';
-var SECRET_FIELDS = ['modem_password', 'mqtt_password', 'admin_password', 'speedtest_tracker_token', 'notify_webhook_token', 'notify_apprise_key', 'notify_apprise_token', 'notify_pwa_push_vapid_private_key']
-    .concat(typeof MODULE_SECRET_FIELDS !== 'undefined' ? MODULE_SECRET_FIELDS : []);
-var SAVED_SECRET_FIELDS = typeof SAVED_MODULE_SECRET_FIELDS !== 'undefined'
-    ? SAVED_MODULE_SECRET_FIELDS
-    : [];
-
-function _isConfigSecretField(inp) {
-    return !!(
-        inp &&
-        inp.name &&
-        inp.tagName === 'INPUT' &&
-        (
-            SECRET_FIELDS.indexOf(inp.name) !== -1 ||
-            inp.dataset.configSecret === 'true' ||
-            _isSavedSecretField(inp)
-        )
-    );
-}
-
-function getFormData() {
-    var form = document.getElementById('settings-form');
-    var data = {};
-    form.querySelectorAll('input:not(#theme-toggle-appearance):not(#isp_other_input):not(.notify-toggle):not(.notify-cooldown-input):not(.module-toggle-input), select:not(#isp_select), textarea').forEach(function(inp) {
-        if (inp.type === 'checkbox') {
-            data[inp.name] = inp.checked ? inp.value : 'false';
-            return;
-        }
-        if (inp.type === 'hidden' && data[inp.name] !== undefined) return;
-        if (_isConfigSecretField(inp)) {
-            if (_isSavedSecretField(inp) && !inp.dataset.userEditedSecret) {
-                data[inp.name] = MASK;
-            } else {
-                data[inp.name] = inp.value || MASK;
-            }
-        } else {
-            data[inp.name] = inp.value;
-        }
-    });
-    var ispSel = document.getElementById('isp_select');
-    if (ispSel.value === '__other__') {
-        data.isp_name = document.getElementById('isp_other_input').value;
-    } else {
-        data.isp_name = ispSel.value;
-    }
-    data.theme = document.documentElement.getAttribute('data-theme') || 'dark';
-    var cooldowns = {};
-    document.querySelectorAll('.notify-event-row').forEach(function(row) {
-        var eventKey = row.getAttribute('data-event');
-        var severity = row.getAttribute('data-severity') || '';
-        var key = severity ? eventKey + ':' + severity : eventKey;
-        var toggle = row.querySelector('.notify-toggle');
-        var inp = row.querySelector('.notify-cooldown-input');
-        if (!toggle.checked) {
-            cooldowns[key] = 0;
-        } else if (inp.value.trim() !== '') {
-            cooldowns[key] = parseInt(inp.value, 10) || 1;
-        }
-    });
-    data.notify_cooldowns = JSON.stringify(cooldowns);
-    return data;
-}
-
-/* ── Modem Test ── */
-function testModem() {
-    var el = document.getElementById('modem-test');
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + T.testing));
-    setModemStatus('testing', T.testing);
-    var data = getFormData();
-    fetch(docsightUrl('/api/test-modem'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify({modem_type: data.modem_type, modem_url: data.modem_url, modem_user: data.modem_user, modem_password: data.modem_password})
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        el.textContent = '';
-        if (res.success) {
-            el.className = 'test-result test-ok';
-            var check = document.createElement('span');
-            check.className = 'check-icon';
-            check.textContent = '\u2713';
-            el.appendChild(check);
-            el.appendChild(document.createTextNode(' ' + T.connected + ': ' + (res.model || 'OK')));
-            setModemStatus('connected', T.connected + ': ' + (res.model || 'OK'));
-        } else {
-            el.className = 'test-result test-fail';
-            var x = document.createElement('span');
-            x.textContent = '\u2717';
-            el.appendChild(x);
-            el.appendChild(document.createTextNode(' ' + T.error_prefix + ': ' + (res.error || T.unknown_error)));
-            setModemStatus('disconnected', T.error_prefix + ': ' + (res.error || T.unknown_error));
-        }
-    })
-    .catch(function() {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + T.network_error));
-        setModemStatus('disconnected', T.network_error);
-    });
-}
-
-/* ── MQTT Test ── */
-function testMqtt() {
-    var el = document.getElementById('mqtt-test');
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + T.testing));
-    setMqttStatus('testing', T.testing);
-    var data = getFormData();
-    fetch(docsightUrl('/api/test-mqtt'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify({mqtt_host: data.mqtt_host, mqtt_port: data.mqtt_port, mqtt_user: data.mqtt_user, mqtt_password: data.mqtt_password})
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        el.textContent = '';
-        if (res.success) {
-            el.className = 'test-result test-ok';
-            var check = document.createElement('span');
-            check.className = 'check-icon';
-            check.textContent = '\u2713';
-            el.appendChild(check);
-            el.appendChild(document.createTextNode(' ' + T.connected));
-            setMqttStatus('connected', T.connected);
-        } else {
-            el.className = 'test-result test-fail';
-            var x = document.createElement('span');
-            x.textContent = '\u2717';
-            el.appendChild(x);
-            el.appendChild(document.createTextNode(' ' + T.error_prefix + ': ' + (res.error || T.unknown_error)));
-            setMqttStatus('disconnected', T.error_prefix + ': ' + (res.error || T.unknown_error));
-        }
-    })
-    .catch(function() {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + T.network_error));
-        setMqttStatus('disconnected', T.network_error);
-    });
-}
-
-/* ── Speedtest Tracker Test ── */
-function testSpeedtest() {
-    var el = document.getElementById('speedtest-test');
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + T.testing));
-    var data = getFormData();
-    fetch(docsightUrl('/api/test-speedtest'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify({
-            speedtest_tracker_url: data.speedtest_tracker_url,
-            speedtest_tracker_token: data.speedtest_tracker_token,
-            speedtest_tls_insecure: data.speedtest_tls_insecure
-        })
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        el.textContent = '';
-        if (res.success) {
-            el.className = 'test-result test-ok';
-            var check = document.createElement('span');
-            check.className = 'check-icon';
-            check.textContent = '\u2713';
-            el.appendChild(check);
-            if (res.results > 0) {
-                el.appendChild(document.createTextNode(' ' + T.connected + ': ' + res.latest.download + ' \u2193 / ' + res.latest.upload + ' \u2191 / ' + res.latest.ping));
-            } else {
-                el.appendChild(document.createTextNode(' ' + T.connected + ' (' + (T.speedtest_no_results || 'no results yet') + ')'));
-            }
-        } else {
-            el.className = 'test-result test-fail';
-            var x = document.createElement('span');
-            x.textContent = '\u2717';
-            el.appendChild(x);
-            el.appendChild(document.createTextNode(' ' + T.error_prefix + ': ' + (res.error || T.unknown_error)));
-        }
-    })
-    .catch(function() {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + T.network_error));
-    });
-}
-
-/* ── Speedtest Cache Clear ── */
-function clearSpeedtestCache(btn) {
-    var message = T.clear_cache_confirm || 'Clear all cached speedtest results? They will be re-synced on the next poll cycle.';
-    docsightConfirm({
-        title: T['docsight.speedtest.clear_cache'] || 'Clear Cache',
-        message: message,
-        confirmText: T['docsight.speedtest.clear_cache'] || 'Clear Cache',
-        cancelText: T.cancel || 'Cancel',
-        danger: true
-    }).then(function(confirmed) {
-        if (!confirmed) return null;
-        btn.disabled = true;
-        return fetch(docsightUrl('/api/speedtest/cache'), { method: 'DELETE' });
-    })
-        .then(function(r) { return r ? r.json() : null; })
-        .then(function(res) {
-            if (!res) return;
-            btn.disabled = false;
-            if (res.success) {
-                var count = res.cleared || 0;
-                btn.textContent = '\u2713 ' + count + ' ' + (T.cleared || 'cleared');
-                setTimeout(function() {
-                    btn.textContent = '';
-                    var icon = document.createElement('i');
-                    icon.setAttribute('data-lucide', 'trash-2');
-                    btn.appendChild(icon);
-                    btn.appendChild(document.createTextNode(' ' + (T['docsight.speedtest.clear_cache'] || 'Clear Cache')));
-                    if (typeof lucide !== 'undefined') lucide.createIcons();
-                }, 2000);
-            }
-        })
-        .catch(function() { btn.disabled = false; });
-}
-
-/* ── Notification Test ── */
-function testNotifications() {
-    var el = document.getElementById('notify-test');
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + (T.notify_test_sending || 'Sending test notification...')));
-    fetch(docsightUrl('/api/notifications/test'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: '{}'
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        el.textContent = '';
-        if (res.success) {
-            el.className = 'test-result test-ok';
-            var check = document.createElement('span');
-            check.className = 'check-icon';
-            check.textContent = '\u2713';
-            el.appendChild(check);
-            el.appendChild(document.createTextNode(' ' + (T.notify_test_sent || 'Test notification sent')));
-        } else {
-            el.className = 'test-result test-fail';
-            var x = document.createElement('span');
-            x.textContent = '\u2717';
-            el.appendChild(x);
-            el.appendChild(document.createTextNode(' ' + (res.error || T.unknown_error || 'Unknown error')));
-        }
-    })
-    .catch(function() {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + (T.network_error || 'Network error')));
-    });
-}
-
-/* ── PWA Web Push ── */
-var _pwaPushStatus = null;
-
-function _setPwaPushStatus(message, ok) {
-    var el = document.getElementById('pwa-push-status');
-    if (!el) return;
-    el.className = 'test-result ' + (ok ? 'test-ok' : 'test-fail');
-    el.style.display = 'flex';
-    el.textContent = message;
-}
-
-function _urlBase64ToUint8Array(base64String) {
-    var padding = '='.repeat((4 - base64String.length % 4) % 4);
-    var base64 = (base64String + padding).replace(/-/g, '+').replace(/_/g, '/');
-    var rawData = atob(base64);
-    var outputArray = new Uint8Array(rawData.length);
-    for (var i = 0; i < rawData.length; ++i) outputArray[i] = rawData.charCodeAt(i);
-    return outputArray;
-}
-
-function _getCurrentPwaSubscription() {
-    if (!('serviceWorker' in navigator) || !('PushManager' in window)) return Promise.resolve(null);
-    return navigator.serviceWorker.ready
-        .then(function(registration) { return registration.pushManager.getSubscription(); })
-        .catch(function() { return null; });
-}
-
-function refreshPwaPushStatus() {
-    if (!document.getElementById('pwa-push-status')) return Promise.resolve(null);
-    if (!('serviceWorker' in navigator) || !('PushManager' in window) || !('Notification' in window)) {
-        _setPwaPushStatus(T.notify_pwa_push_unsupported || 'Web Push is not supported in this browser.', false);
-        return Promise.resolve(null);
-    }
-    return Promise.all([
-        fetch(docsightUrl('/api/notifications/pwa/status')).then(function(r) { return r.json(); }),
-        _getCurrentPwaSubscription()
-    ])
-        .then(function(values) {
-            var status = values[0];
-            var currentSubscription = values[1];
-            _pwaPushStatus = status;
-            var permission = Notification.permission;
-            if (!status.configured) {
-                _setPwaPushStatus(T.notify_pwa_push_not_configured || 'Web Push is not configured yet.', false);
-            } else if (permission === 'granted' && currentSubscription) {
-                _setPwaPushStatus((T.notify_pwa_push_ready || 'This browser is subscribed.') + ' ' + (status.subscription_count || 0) + ' ' + (T.notify_pwa_push_devices || 'device(s)'), true);
-            } else if (permission === 'granted') {
-                _setPwaPushStatus(T.notify_pwa_push_not_subscribed || 'This browser is not subscribed yet.', false);
-            } else {
-                _setPwaPushStatus(T.notify_pwa_push_permission_needed || 'Browser permission is needed for this device.', false);
-            }
-            return status;
-        })
-        .catch(function() {
-            _setPwaPushStatus(T.network_error || 'Network error', false);
-            return null;
-        });
-}
-
-function subscribePwaPush() {
-    refreshPwaPushStatus().then(function(status) {
-        if (!status || !status.configured || !status.public_key) return;
-        return navigator.serviceWorker.ready
-            .then(function(registration) { return Notification.requestPermission().then(function(permission) { return {registration: registration, permission: permission}; }); })
-            .then(function(ctx) {
-                if (ctx.permission !== 'granted') throw new Error(T.notify_pwa_push_permission_denied || 'Notification permission was denied.');
-                return ctx.registration.pushManager.getSubscription().then(function(existing) {
-                    if (existing) return existing;
-                    return ctx.registration.pushManager.subscribe({
-                        userVisibleOnly: true,
-                        applicationServerKey: _urlBase64ToUint8Array(status.public_key)
-                    });
-                });
-            })
-            .then(function(subscription) {
-                return fetch(docsightUrl('/api/notifications/pwa/subscribe'), {
-                    method: 'POST',
-                    headers: {'Content-Type': 'application/json'},
-                    body: JSON.stringify({subscription: subscription.toJSON()})
-                });
-            })
-            .then(function(r) { return r.json(); })
-            .then(function(res) {
-                if (!res.success) throw new Error(res.error || T.unknown_error || 'Unknown error');
-                showToast(T.notify_pwa_push_subscribed || 'This browser is subscribed.', true);
-                refreshPwaPushStatus();
-            })
-            .catch(function(err) {
-                showToast(err.message || (T.notify_pwa_push_subscribe_failed || 'Subscription failed'), false);
-            });
-    });
-}
-
-function unsubscribePwaPush() {
-    if (!('serviceWorker' in navigator)) return;
-    navigator.serviceWorker.ready
-        .then(function(registration) { return registration.pushManager.getSubscription(); })
-        .then(function(subscription) {
-            if (!subscription) return null;
-            var endpoint = subscription.endpoint;
-            return subscription.unsubscribe().then(function() {
-                return fetch(docsightUrl('/api/notifications/pwa/unsubscribe'), {
-                    method: 'POST',
-                    headers: {'Content-Type': 'application/json'},
-                    body: JSON.stringify({endpoint: endpoint})
-                });
-            });
-        })
-        .then(function(r) { return r ? r.json() : {success: true}; })
-        .then(function(res) {
-            if (!res.success) throw new Error(res.error || T.unknown_error || 'Unknown error');
-            showToast(T.notify_pwa_push_unsubscribed || 'This browser is unsubscribed.', true);
-            refreshPwaPushStatus();
-        })
-        .catch(function(err) {
-            showToast(err.message || (T.notify_pwa_push_unsubscribe_failed || 'Unsubscribe failed'), false);
-        });
-}
-
-/* ── Unsaved Changes ── */
-var _formDirty = false;
-var _manualDirty = false;
-var _formBaseline = null;
-var _manualFormBaseline = null;
-var _instantSaveFailedDirty = false;
-var _secretEditVersionCounter = 0;
-
-function _isSavedSecretField(el) {
-    return !!(
-        el &&
-        el.name &&
-        el.tagName === 'INPUT' &&
-        el.type === 'password' &&
-        (
-            el.dataset.savedSecret === 'true' ||
-            SAVED_SECRET_FIELDS.indexOf(el.name) !== -1
-        )
-    );
-}
-
-function _shouldTreatSavedSecretEventAsUserEdit(e) {
-    var target = e && e.target;
-    return !!(
-        e &&
-        e.isTrusted &&
-        _isSavedSecretField(target) &&
-        document.activeElement === target
-    );
-}
-
-function _normalizedFormValue(el) {
-    if (_isSavedSecretField(el)) {
-        return el.dataset.userEditedSecret ? '__edited_secret__' : '';
-    }
-    if (el.type === 'checkbox' || el.type === 'radio') {
-        return el.checked ? '1' : '0';
-    }
-    if (el.tagName === 'SELECT' && el.multiple) {
-        return Array.prototype.map.call(el.selectedOptions, function(opt) { return opt.value; }).sort().join('\u0000');
-    }
-    return el.value || '';
-}
-
-function _serializeSettingsForm(form, options) {
-    if (!form) return '[]';
-    options = options || {};
-    var ignoreInstantSaveControls = options.ignoreInstantSaveControls === true;
-    var fields = [];
-    Array.prototype.forEach.call(form.elements, function(el) {
-        if (!el.name || el.type === 'submit' || el.type === 'button' || el.type === 'file') return;
-        if (ignoreInstantSaveControls && _isInstantSaveControl(el)) return;
-        var field = { name: el.name, id: el.getAttribute('data-module-id') || '', type: el.type || el.tagName, value: _normalizedFormValue(el) };
-        if (_isSavedSecretField(el) && el.dataset.userEditedSecret) {
-            field.secretEditVersion = el.dataset.secretEditVersion || '';
-        }
-        fields.push(field);
-    });
-    if (!ignoreInstantSaveControls) document.querySelectorAll('.notify-event-row').forEach(function(row) {
-        var eventKey = row.getAttribute('data-event');
-        var severity = row.getAttribute('data-severity') || '';
-        var key = severity ? eventKey + ':' + severity : eventKey;
-        var toggle = row.querySelector('.notify-toggle');
-        var inp = row.querySelector('.notify-cooldown-input');
-        if (!eventKey || !toggle || !inp) return;
-        fields.push({
-            name: 'notify_cooldown:' + key,
-            id: key,
-            type: 'notification-cooldown',
-            value: toggle.checked ? (inp.value || '') : '0'
-        });
-    });
-    fields.sort(function(a, b) {
-        return (a.name + ':' + a.id + ':' + a.type).localeCompare(b.name + ':' + b.id + ':' + b.type);
-    });
-    return JSON.stringify(fields);
-}
-
-function _serializeManualSettingsForm(form) {
-    return _serializeSettingsForm(form, { ignoreInstantSaveControls: true });
-}
-
-function _isInstantSaveFieldRecord(field, form) {
-    if (!field || !form) return false;
-    if (field.type === 'notification-cooldown') return true;
-    return Array.prototype.some.call(form.elements, function(el) {
-        if (!el || el.name !== field.name || !_isInstantSaveControl(el)) return false;
-        var moduleId = el.getAttribute('data-module-id') || '';
-        return (field.id || '') === moduleId && (el.type || el.tagName) === field.type;
-    });
-}
-
-function _manualBaselineFromSerialized(serializedBaseline, form) {
-    if (!form) return '[]';
-    try {
-        var fields = JSON.parse(serializedBaseline || '[]').filter(function(field) {
-            return !_isInstantSaveFieldRecord(field, form);
-        });
-        fields.sort(function(a, b) {
-            return (a.name + ':' + a.id + ':' + a.type).localeCompare(b.name + ':' + b.id + ':' + b.type);
-        });
-        return JSON.stringify(fields);
-    } catch(e) {
-        return _serializeManualSettingsForm(form);
-    }
-}
-
-function _captureFormBaseline() {
-    var form = document.getElementById('settings-form');
-    _formBaseline = _serializeSettingsForm(form);
-    _manualFormBaseline = _serializeManualSettingsForm(form);
-}
-
-function _hasManualDirtyChanges() {
-    var form = document.getElementById('settings-form');
-    if (!form) return false;
-    if (_manualFormBaseline === null) _captureFormBaseline();
-    return _serializeManualSettingsForm(form) !== _manualFormBaseline;
-}
-
-function _syncSaveFooter() {
-    var footer = document.getElementById('save-footer');
-    if (!footer) return;
-    var shouldShowFooter = _formDirty && !_saveHiddenSections[_currentSection] && (_instantSaveFailedDirty || _manualDirty);
-    if (shouldShowFooter) {
-        footer.classList.add('visible');
-        footer.removeAttribute('aria-hidden');
-        footer.removeAttribute('inert');
-    } else {
-        footer.classList.remove('visible');
-        footer.setAttribute('aria-hidden', 'true');
-        footer.setAttribute('inert', '');
-    }
-}
-
-function _updateDirtyState(options) {
-    options = options || {};
-    var form = document.getElementById('settings-form');
-    if (!form) return;
-    if (_formBaseline === null) _captureFormBaseline();
-    _formDirty = _serializeSettingsForm(form) !== _formBaseline;
-    if (options.updateManualDirty !== false) {
-        _manualDirty = _hasManualDirtyChanges();
-    }
-    _syncSaveFooter();
-}
-
-function _isInstantSaveControl(el) {
-    if (!el || !el.matches) return false;
-    if (el.id === 'theme-toggle-appearance') return false;
-    return el.matches([
-        'label.toggle input[type="checkbox"]',
-        'label.switch input[type="checkbox"]',
-        '.module-toggle-input',
-        '.notify-toggle',
-        '.notify-cooldown-input'
-    ].join(','));
-}
-
-function _isInstantSaveDirtyEvent(e) {
-    return _isInstantSaveControl(e && e.target);
-}
-
-function initUnsavedDetection() {
-    var form = document.getElementById('settings-form');
-    if (!form) return;
-    _captureFormBaseline();
-    function markDirty(e) {
-        if (_shouldTreatSavedSecretEventAsUserEdit(e)) {
-            e.target.dataset.userEditedSecret = 'true';
-            e.target.dataset.secretEditVersion = String(++_secretEditVersionCounter);
-        }
-        if (_isInstantSaveDirtyEvent(e)) {
-            _updateDirtyState({ updateManualDirty: false });
-            return;
-        }
-        _updateDirtyState();
-    }
-    form.addEventListener('input', markDirty);
-    form.addEventListener('change', markDirty);
-
-    /* P1: warn on page leave when unsaved changes exist */
-    window.addEventListener('beforeunload', function(e) {
-        if (_formDirty) {
-            e.preventDefault();
-            e.returnValue = '';
-        }
-    });
-}
-
-function _resetEditedSavedSecrets(savedSecretVersions) {
-    var form = document.getElementById('settings-form');
-    if (!form) return;
-    Array.prototype.forEach.call(form.elements, function(el) {
-        if (!_isSavedSecretField(el) || !el.dataset.userEditedSecret) return;
-        if (savedSecretVersions && savedSecretVersions[el.name] !== el.dataset.secretEditVersion) return;
-        el.value = '';
-        el.dataset.savedSecret = 'true';
-        delete el.dataset.userEditedSecret;
-        delete el.dataset.secretEditVersion;
-    });
-}
-
-function clearDirty() {
-    _resetEditedSavedSecrets();
-    _captureFormBaseline();
-    _formDirty = false;
-    _manualDirty = false;
-    _instantSaveFailedDirty = false;
-    _syncSaveFooter();
-}
-
-function _getPendingModuleStates() {
-    var states = [];
-    document.querySelectorAll('.module-toggle-input').forEach(function(toggle) {
-        var moduleId = toggle.getAttribute('data-module-id');
-        if (!moduleId) return;
-        var initial = toggle.getAttribute('data-initial-enabled') === 'true';
-        if (toggle.checked !== initial) {
-            states.push({ id: moduleId, enabled: toggle.checked });
-        }
-    });
-    return states;
-}
-
-function _markModuleStatesSaved(states) {
-    var savedById = null;
-    if (states) {
-        savedById = {};
-        states.forEach(function(state) {
-            savedById[state.id] = state.enabled;
-        });
-    }
-    document.querySelectorAll('.module-toggle-input').forEach(function(toggle) {
-        var moduleId = toggle.getAttribute('data-module-id');
-        if (savedById && savedById[moduleId] === undefined) return;
-        var savedEnabled = savedById ? savedById[moduleId] : toggle.checked;
-        toggle.setAttribute('data-initial-enabled', savedEnabled ? 'true' : 'false');
-    });
-}
-
-function _saveModuleStatesIfNeeded(states) {
-    states = states || _getPendingModuleStates();
-    if (states.length === 0) return Promise.resolve({ success: true, restart_required: false });
-    return fetch(docsightUrl('/api/modules/batch'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify({ modules: states })
-    })
-    .then(function(r) {
-        return r.json().then(function(data) { return { ok: r.ok, data: data }; });
-    })
-    .then(function(res) {
-        if (!res.ok || !res.data.success) {
-            throw new Error(res.data.error || T.save_failed || 'Save failed');
-        }
-        _markModuleStatesSaved(states);
-        if (res.data.restart_required) {
-            var banner = document.getElementById('module-restart-banner');
-            if (banner) {
-                banner.style.display = 'flex';
-                if (typeof lucide !== 'undefined') lucide.createIcons({nodes: [banner]});
-            }
-        }
-        return res.data;
-    });
-}
-
-function _savedBaselineAfterSecretReset(serializedBaseline) {
-    try {
-        var fields = JSON.parse(serializedBaseline || '[]');
-        fields.forEach(function(field) {
-            if (field && field.value === '__edited_secret__') {
-                field.value = '';
-                delete field.secretEditVersion;
-            }
-        });
-        return JSON.stringify(fields);
-    } catch(e) {
-        return serializedBaseline;
-    }
-}
-
-function _savedSecretVersionsFromBaseline(serializedBaseline) {
-    var versions = {};
-    try {
-        var fields = JSON.parse(serializedBaseline || '[]');
-        fields.forEach(function(field) {
-            if (field && field.value === '__edited_secret__') {
-                versions[field.name] = field.secretEditVersion || '';
-            }
-        });
-    } catch(e) {}
-    return versions;
-}
-
-function _finishSettingsSave(savedFormBaseline, options) {
-    options = options || {};
-    var showSuccessToast = options.showSuccessToast !== false;
-    var form = document.getElementById('settings-form');
-    if (!savedFormBaseline || !form) {
-        clearDirty();
-    } else {
-        var savedBaseline = _savedBaselineAfterSecretReset(savedFormBaseline);
-        var savedSecretVersions = _savedSecretVersionsFromBaseline(savedFormBaseline);
-        _resetEditedSavedSecrets(savedSecretVersions);
-        if (_serializeSettingsForm(form) === savedBaseline) {
-            _formBaseline = savedBaseline;
-            _manualFormBaseline = _manualBaselineFromSerialized(savedBaseline, form);
-            _instantSaveFailedDirty = false;
-            _manualDirty = false;
-            _formDirty = false;
-            _syncSaveFooter();
-        } else {
-            _formBaseline = savedBaseline;
-            _manualFormBaseline = _manualBaselineFromSerialized(savedBaseline, form);
-            _instantSaveFailedDirty = false;
-            _updateDirtyState({ updateManualDirty: false });
-        }
-    }
-    if (showSuccessToast) showToast(T.settings_saved || 'Settings saved', true);
-    var newLang = document.getElementById('language').value;
-    var newTz = document.getElementById('timezone').value;
-    if (newLang !== currentLang || newTz !== currentTz) {
-        setTimeout(function() { location.reload(); }, 800);
-    }
-}
-
-function _saveAllSettings(options) {
-    var form = document.getElementById('settings-form');
-    var data = getFormData();
-    var savedFormBaseline = _serializeSettingsForm(form);
-    var moduleStates = _getPendingModuleStates();
-    return fetch(docsightUrl('/api/config'), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify(data)
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        if (!res.success) {
-            throw new Error(res.error || T.save_failed || 'Save failed');
-        }
-        return _saveModuleStatesIfNeeded(moduleStates);
-    })
-    .then(function() {
-        _finishSettingsSave(savedFormBaseline, options);
-        return true;
-    });
-}
-
-/**
- * Save the settings form via /api/config and pending module state.
- * Uses the same error display and lang/tz reload logic as the normal submit.
- * Returns a Promise that resolves to true on success, false on failure.
- */
-function _saveForm() {
-    var errEl = document.getElementById('global-error');
-    if (errEl) errEl.style.display = 'none';
-    return _enqueueSettingsSave()
-    .catch(function(err) {
-        if (errEl) {
-            errEl.textContent = err.message || T.network_error;
-            errEl.style.display = 'block';
-        }
-        return false;
-    });
-}
-
-var _settingsSaveQueue = Promise.resolve();
-
-function _enqueueSettingsSave(options) {
-    _settingsSaveQueue = _settingsSaveQueue.catch(function() { return true; }).then(function() {
-        return _saveAllSettings(options);
-    });
-    return _settingsSaveQueue;
-}
-
-function _saveSettingsInstantly() {
-    var errEl = document.getElementById('global-error');
-    if (errEl) errEl.style.display = 'none';
-    return _enqueueSettingsSave({ showSuccessToast: false }).catch(function(err) {
-        if (errEl) {
-            errEl.textContent = err.message || T.network_error;
-            errEl.style.display = 'block';
-        }
-        _instantSaveFailedDirty = true;
-        _updateDirtyState({ updateManualDirty: false });
-        return false;
-    });
-}
-
-/**
- * If the form has unsaved changes, ask the user to save or discard.
- * OK = save first, then proceed.  Cancel = abort the action entirely.
- * Returns a Promise that resolves to true if the action may proceed.
- */
-function _guardUnsaved() {
-    if (!_formDirty) return Promise.resolve(true);
-    var msg = T.unsaved_confirm || 'You have unsaved settings changes. Save them before continuing?';
-    return docsightConfirm({
-        title: T.unsaved_changes || 'Unsaved changes',
-        message: msg,
-        confirmText: T.save || 'Save',
-        cancelText: T.cancel || 'Cancel'
-    }).then(function(confirmed) {
-        if (confirmed) return _saveForm();
-        /* User chose Cancel - abort the action, keep unsaved changes */
-        return false;
-    });
-}
-
-/* ── Submit ── */
-function initFormSubmit() {
-    document.getElementById('settings-form').addEventListener('submit', function(e) {
-        e.preventDefault();
-        var errEl = document.getElementById('global-error');
-        errEl.style.display = 'none';
-        _enqueueSettingsSave()
-        .catch(function(err) {
-            errEl.textContent = err.message || T.network_error;
-            errEl.style.display = 'block';
-        });
-    });
-}
-
-/* ── Timezone Hint ── */
-function initTimezoneHint() {
-    if (typeof serverOffsetMin === 'undefined') return;
-    var browserOffsetMin = -new Date().getTimezoneOffset();
-    var diffMin = browserOffsetMin - serverOffsetMin;
-    if (diffMin === 0) return;
-
-    var inp = document.getElementById('snapshot_time');
-    var hint = document.getElementById('snapshot-tz-hint');
-    if (!inp || !hint) return;
-
-    function updateTzHint() {
-        var parts = inp.value.split(':');
-        if (parts.length < 2) return;
-        var h = parseInt(parts[0]), m = parseInt(parts[1]);
-        var totalMin = h * 60 + m + diffMin;
-        totalMin = ((totalMin % 1440) + 1440) % 1440;
-        var lh = String(Math.floor(totalMin / 60)).padStart(2, '0');
-        var lm = String(totalMin % 60).padStart(2, '0');
-        hint.textContent = T.snapshot_hint + ' \u2014 ' + inp.value + ' ' + serverTz + ' = ' + lh + ':' + lm + ' ' + T.snapshot_your_time;
-    }
-    updateTzHint();
-    inp.addEventListener('change', updateTzHint);
-}
-
-/* ── Backup ── */
-function downloadBackup() {
-    var btn = document.getElementById('backup-download-btn');
-    var el = document.getElementById('backup-download-result');
-    btn.disabled = true;
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + (T.backup_creating || 'Creating backup...')));
-    fetch(docsightUrl('/api/backup'), { method: 'POST' })
-    .then(function(r) {
-        if (!r.ok) return r.json().then(function(j) { throw new Error(j.error || 'Backup failed'); });
-        var cd = r.headers.get('Content-Disposition') || '';
-        var match = cd.match(/filename="?([^"]+)"?/);
-        var fname = match ? match[1] : 'docsight_backup.tar.gz';
-        return r.blob().then(function(blob) { return { blob: blob, fname: fname }; });
-    })
-    .then(function(res) {
-        var a = document.createElement('a');
-        a.href = URL.createObjectURL(res.blob);
-        a.download = res.fname;
-        document.body.appendChild(a);
-        a.click();
-        document.body.removeChild(a);
-        URL.revokeObjectURL(a.href);
-        el.textContent = '';
-        el.className = 'test-result test-ok';
-        var check = document.createElement('span');
-        check.className = 'check-icon';
-        check.textContent = '\u2713';
-        el.appendChild(check);
-        el.appendChild(document.createTextNode(' ' + (T.backup_success || 'Backup downloaded')));
-        btn.disabled = false;
-    })
-    .catch(function(err) {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + (err.message || T.network_error)));
-        btn.disabled = false;
-    });
-}
-
-function backupNow() {
-    var el = document.getElementById('backup-now-result');
-    el.className = 'test-result test-loading';
-    el.style.display = 'flex';
-    el.textContent = '';
-    var span = document.createElement('span');
-    span.textContent = '\u23F3';
-    el.appendChild(span);
-    el.appendChild(document.createTextNode(' ' + (T.backup_creating || 'Creating backup...')));
-    fetch(docsightUrl('/api/backup/scheduled'), { method: 'POST' })
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        el.textContent = '';
-        if (res.success) {
-            el.className = 'test-result test-ok';
-            var check = document.createElement('span');
-            check.className = 'check-icon';
-            check.textContent = '\u2713';
-            el.appendChild(check);
-            el.appendChild(document.createTextNode(' ' + (T.backup_saved || 'Backup saved') + ': ' + res.filename));
-            loadBackupList();
-        } else {
-            el.className = 'test-result test-fail';
-            var x = document.createElement('span');
-            x.textContent = '\u2717';
-            el.appendChild(x);
-            el.appendChild(document.createTextNode(' ' + (res.error || T.save_failed || 'Failed')));
-        }
-    })
-    .catch(function() {
-        el.textContent = '';
-        el.className = 'test-result test-fail';
-        var x = document.createElement('span');
-        x.textContent = '\u2717';
-        el.appendChild(x);
-        el.appendChild(document.createTextNode(' ' + (T.network_error || 'Network error')));
-    });
-}
-
-function loadBackupList() {
-    var el = document.getElementById('backup-list');
-    if (!el) return;
-    fetch(docsightUrl('/api/backup/list'))
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        var backups = Array.isArray(res) ? res : (res && Array.isArray(res.backups) ? res.backups : []);
-        el.textContent = '';
-        if (backups.length === 0) {
-            var emptySpan = document.createElement('span');
-            emptySpan.style.cssText = 'color:var(--muted);font-style:italic;';
-            emptySpan.textContent = T.backup_none || 'No backups found';
-            el.appendChild(emptySpan);
-            return;
-        }
-        var table = document.createElement('table');
-        table.style.cssText = 'width:100%;border-collapse:collapse;';
-        var tbody = document.createElement('tbody');
-        backups.forEach(function(b) {
-            var sizeMB = (b.size / 1048576).toFixed(1);
-            var date = b.modified ? new Date(b.modified).toLocaleString() : '';
-            var tr = document.createElement('tr');
-            tr.style.cssText = 'border-bottom:1px solid var(--card-border);';
-
-            var td1 = document.createElement('td');
-            td1.style.cssText = 'padding:6px 0;';
-            var codeEl = document.createElement('code');
-            codeEl.style.cssText = 'font-size:0.8em;';
-            codeEl.textContent = b.filename;
-            td1.appendChild(codeEl);
-            tr.appendChild(td1);
-
-            var td2 = document.createElement('td');
-            td2.style.cssText = 'padding:6px 8px;color:var(--muted);font-size:0.8em;white-space:nowrap;';
-            td2.textContent = date;
-            tr.appendChild(td2);
-
-            var td3 = document.createElement('td');
-            td3.style.cssText = 'padding:6px 8px;color:var(--muted);font-size:0.8em;white-space:nowrap;';
-            td3.textContent = sizeMB + ' MB';
-            tr.appendChild(td3);
-
-            var td4 = document.createElement('td');
-            td4.style.cssText = 'padding:6px 0;text-align:right;';
-            var delBtn = document.createElement('button');
-            delBtn.type = 'button';
-            delBtn.className = 'btn btn-secondary';
-            delBtn.style.cssText = 'padding:2px 8px;font-size:0.75em;';
-            var deleteLabel = ((T.backup_delete || T.bqm_delete || 'Delete backup') + ' ' + b.filename).trim();
-            delBtn.setAttribute('data-filename', b.filename);
-            delBtn.setAttribute('aria-label', deleteLabel);
-            delBtn.setAttribute('title', deleteLabel);
-            delBtn.addEventListener('click', function() { deleteBackup(this.getAttribute('data-filename')); });
-            var icon = document.createElement('i');
-            icon.setAttribute('data-lucide', 'trash-2');
-            icon.setAttribute('aria-hidden', 'true');
-            icon.style.cssText = 'width:12px;height:12px;';
-            delBtn.appendChild(icon);
-            td4.appendChild(delBtn);
-            tr.appendChild(td4);
-
-            tbody.appendChild(tr);
-        });
-        table.appendChild(tbody);
-        el.appendChild(table);
-        if (typeof lucide !== 'undefined') lucide.createIcons();
-    })
-    .catch(function() {
-        el.textContent = '';
-        var errSpan = document.createElement('span');
-        errSpan.style.cssText = 'color:var(--error);';
-        errSpan.textContent = T.network_error;
-        el.appendChild(errSpan);
-    });
-}
-
-function deleteBackup(filename) {
-    docsightConfirm({
-        title: T.delete || 'Delete',
-        message: T.backup_delete_confirm || 'Delete this backup?',
-        confirmText: T.delete || 'Delete',
-        cancelText: T.cancel || 'Cancel',
-        danger: true
-    }).then(function(confirmed) {
-        if (!confirmed) return null;
-        return fetch(docsightUrl('/api/backup/' + encodeURIComponent(filename)), { method: 'DELETE' });
-    })
-    .then(function(r) { return r ? r.json() : null; })
-    .then(function(res) {
-        if (!res) return;
-        if (res.success) {
-            loadBackupList();
-        } else {
-            showToast(res.error || T.save_failed || 'Failed', false);
-        }
-    })
-    .catch(function() { showToast(T.network_error || 'Network error', false); });
-}
-
-/* ── Directory Browser ── */
-var _browsePath = '/backup';
-
-function openBrowseModal(opener) {
-    var modal = document.getElementById('browse-modal');
-    _browsePath = document.getElementById('backup_path').value || '/backup';
-    var selected = document.getElementById('browse-selected-path');
-    var status = document.getElementById('browse-status');
-    if (selected) selected.textContent = _browsePath;
-    if (status) status.textContent = T.loading || 'Loading...';
-    DOCSightModal.open(modal, {opener: opener || document.activeElement, labelledBy: 'browse-modal-title'});
-    browseTo(_browsePath);
-}
-
-function closeBrowseModal() {
-    DOCSightModal.close('browse-modal');
-}
-
-function selectBrowsePath() {
-    document.getElementById('backup_path').value = _browsePath;
-    closeBrowseModal();
-}
-
-function browseTo(path) {
-    _browsePath = path;
-    var bc = document.getElementById('browse-breadcrumb');
-    var dirs = document.getElementById('browse-dirs');
-    var selected = document.getElementById('browse-selected-path');
-    var status = document.getElementById('browse-status');
-    bc.textContent = path;
-    if (selected) selected.textContent = path;
-    dirs.textContent = '';
-    var loadingDiv = document.createElement('div');
-    loadingDiv.style.cssText = 'padding:16px;color:var(--muted);text-align:center;';
-    loadingDiv.textContent = '\u23F3 ' + (T.loading || 'Loading...');
-    dirs.appendChild(loadingDiv);
-    if (status) status.textContent = T.loading || 'Loading...';
-    fetch(docsightUrl('/api/browse?path=' + encodeURIComponent(path)))
-    .then(function(r) { return r.json(); })
-    .then(function(res) {
-        dirs.textContent = '';
-        if (res.error) {
-            var errDiv = document.createElement('div');
-            errDiv.style.cssText = 'padding:16px;color:var(--error);';
-            errDiv.textContent = res.error;
-            dirs.appendChild(errDiv);
-            if (status) status.textContent = res.error;
-            return;
-        }
-        _browsePath = res.path;
-        bc.textContent = res.path;
-        if (selected) selected.textContent = res.path;
-
-        if (res.parent) {
-            var parentItem = _createBrowseItem('..', res.parent, 'corner-left-up', true);
-            dirs.appendChild(parentItem);
-        }
-        if (res.directories.length === 0 && !res.parent) {
-            var emptyDiv = document.createElement('div');
-            emptyDiv.style.cssText = 'padding:16px;color:var(--muted);text-align:center;font-style:italic;';
-            emptyDiv.textContent = T.backup_empty_dir || 'Empty directory';
-            dirs.appendChild(emptyDiv);
-        }
-        res.directories.forEach(function(d) {
-            var item = _createBrowseItem(d.name, d.path, 'folder', false);
-            dirs.appendChild(item);
-        });
-        if (status) {
-            status.textContent = res.directories.length
-                ? res.directories.length + ' ' + (T.directories || 'directories')
-                : (T.backup_empty_dir || 'Empty directory');
-        }
-        if (typeof lucide !== 'undefined') lucide.createIcons();
-    })
-    .catch(function() {
-        dirs.textContent = '';
-        var errDiv = document.createElement('div');
-        errDiv.style.cssText = 'padding:16px;color:var(--error);';
-        errDiv.textContent = T.network_error;
-        dirs.appendChild(errDiv);
-        if (status) status.textContent = T.network_error;
-    });
-}
-
-function _createBrowseItem(label, targetPath, iconName, isMuted) {
-    var div = document.createElement('div');
-    div.className = 'browse-item';
-    div.setAttribute('role', 'button');
-    div.setAttribute('tabindex', '0');
-    div.setAttribute('aria-label', label);
-    div.style.cssText = 'padding:8px 12px;cursor:pointer;display:flex;align-items:center;gap:8px;border-radius:var(--radius-sm);';
-    div.addEventListener('click', function() { browseTo(targetPath); });
-    div.addEventListener('keydown', function(event) {
-        if (event.key !== 'Enter' && event.key !== ' ') return;
-        event.preventDefault();
-        browseTo(targetPath);
-    });
-    div.addEventListener('mouseenter', function() { this.style.background = 'var(--hover-bg)'; });
-    div.addEventListener('mouseleave', function() { this.style.background = ''; });
-
-    var icon = document.createElement('i');
-    icon.setAttribute('data-lucide', iconName);
-    icon.style.cssText = 'width:16px;height:16px;color:' + (isMuted ? 'var(--muted)' : 'var(--accent)') + ';';
-    div.appendChild(icon);
-
-    var span = document.createElement('span');
-    span.textContent = label;
-    if (isMuted) span.style.color = 'var(--muted)';
-    div.appendChild(span);
-
-    return div;
-}
-
-/* ── Username Field Toggle + Modem Defaults (data-driven) ── */
-var _previousDriverDefault = '';
-
-function toggleUsernameField() {
-    var modemType = document.getElementById('modem_type');
-    if (!modemType) return;
-
-    var usernameField = document.getElementById('modem_user');
-    var urlField = document.getElementById('modem_url');
-    var passwordField = document.getElementById('modem_password');
-    var hints = (typeof DRIVER_HINTS !== 'undefined' && DRIVER_HINTS[modemType.value]) || {};
-
-    // Fields to hide when credentials not required (e.g. generic driver)
-    var credFields = [urlField, usernameField, passwordField].map(function(f) {
-        return f ? f.closest('.form-field') : null;
-    });
-    var testBtn = document.querySelector('[onclick="testModem()"]');
-    var testBtnParent = testBtn ? testBtn.parentElement : null;
-    var testResult = document.getElementById('modem-test');
-
-    // URL default: apply only if field is empty or shows the previous driver's default
-    if (hints.default_url && urlField && (!urlField.value || urlField.value === _previousDriverDefault)) {
-        urlField.value = hints.default_url;
-    }
-    _previousDriverDefault = hints.default_url || '';
-
-    if (hints.credentials_required === false) {
-        credFields.forEach(function(el) { if (el) el.style.display = 'none'; });
-        if (testBtnParent) testBtnParent.style.display = 'none';
-        if (testResult) testResult.style.display = 'none';
-        return;
-    }
-    credFields.forEach(function(el) { if (el) el.style.display = ''; });
-    if (testBtnParent) testBtnParent.style.display = '';
-
-    // Username handling
-    if (hints.username_required === false) {
-        usernameField.disabled = true;
-        usernameField.value = '';
-        usernameField.placeholder = T.not_required || 'Not required for this modem';
-        usernameField.style.opacity = '0.5';
-        usernameField.style.cursor = 'not-allowed';
-    } else {
-        usernameField.disabled = false;
-        usernameField.style.opacity = '1';
-        usernameField.style.cursor = 'text';
-        if (hints.default_user) {
-            if (!usernameField.value) usernameField.value = hints.default_user;
-            usernameField.placeholder = hints.default_user;
-        } else {
-            usernameField.placeholder = 'admin';
-        }
-    }
-}
-
-function initInstantSettingsToggles() {
-    var form = document.getElementById('settings-form');
-    if (!form) return;
-    var selector = [
-        'label.toggle input[type="checkbox"]:not(#theme-toggle-appearance):not(.module-toggle-input):not(.notify-toggle)',
-        'label.switch input[type="checkbox"]'
-    ].join(',');
-    form.querySelectorAll(selector).forEach(function(toggle) {
-        toggle.addEventListener('change', function() {
-            if (toggle.id === 'font-toggle') {
-                applyFontToggle(toggle.checked);
-            }
-            _saveSettingsInstantly();
-        });
-    });
-}
-
-function initNotificationCooldownControls() {
-    try {
-        var saved = typeof savedCooldowns !== 'undefined' ? savedCooldowns : {};
-        document.querySelectorAll('.notify-event-row').forEach(function(row) {
-            var eventKey = row.getAttribute('data-event');
-            var severity = row.getAttribute('data-severity') || '';
-            var key = severity ? eventKey + ':' + severity : eventKey;
-            var toggle = row.querySelector('.notify-toggle');
-            var inp = row.querySelector('.notify-cooldown-input');
-            if (!toggle || !inp) return;
-            var savedValue = saved[key];
-            if (savedValue === undefined && severity) {
-                savedValue = saved[eventKey];
-            }
-            if (savedValue !== undefined) {
-                if (savedValue === 0) {
-                    toggle.checked = false;
-                    inp.disabled = true;
-                    inp.style.opacity = '0.4';
-                } else {
-                    inp.value = savedValue;
-                }
-            }
-            toggle.addEventListener('change', function() {
-                inp.disabled = !toggle.checked;
-                inp.style.opacity = toggle.checked ? '1' : '0.4';
-                _saveSettingsInstantly();
-            });
-            inp.addEventListener('change', function() {
-                _saveSettingsInstantly();
-            });
-        });
-    } catch(e) {}
-}
-
-/* ── Init ── */
-document.addEventListener('DOMContentLoaded', function() {
-    if (typeof lucide !== 'undefined') lucide.createIcons();
-
-    initThemeToggle();
-    initFormSubmit();
-    initUnsavedDetection();
-    initInstantSettingsToggles();
-    initNotificationCooldownControls();
-    refreshPwaPushStatus();
-    initTimezoneHint();
-    onIspChange();
-    toggleUsernameField();
-    updateNotificationChannelSummaries();
-    syncMobileSidebarAccessibility();
-    if (_mobileSidebarMedia) {
-        if (typeof _mobileSidebarMedia.addEventListener === 'function') {
-            _mobileSidebarMedia.addEventListener('change', syncMobileSidebarAccessibility);
-        } else if (typeof _mobileSidebarMedia.addListener === 'function') {
-            _mobileSidebarMedia.addListener(syncMobileSidebarAccessibility);
-        }
-    }
-
-    /* Listen for input changes on integration fields to update dots and card summaries */
-    ['notify_webhook_url'].forEach(function(id) {
-        var el = document.getElementById(id);
-        if (el) {
-            el.addEventListener('input', updateNotificationChannelSummaries);
-        }
-    });
-    ['notify_apprise_enabled', 'notify_pwa_push_enabled'].forEach(function(id) {
-        var el = document.getElementById(id);
-        if (el) el.addEventListener('change', function() {
-            updateNotificationChannelSummaries();
-            expandNotificationChannelCard(el);
-        });
-    });
-
-    /* Modem type change */
-    var modemType = document.getElementById('modem_type');
-    if (modemType) modemType.addEventListener('change', toggleUsernameField);
-
-    /* Backup toggle */
-    var backupCb = document.getElementById('backup_enabled');
-    var backupSettings = document.getElementById('backup-auto-settings');
-    if (backupCb && backupSettings) {
-        backupCb.addEventListener('change', function() {
-            backupSettings.style.opacity = backupCb.checked ? '1' : '0.5';
-            backupSettings.style.pointerEvents = backupCb.checked ? 'auto' : 'none';
-        });
-    }
-
-    _captureFormBaseline();
-    _formDirty = false;
-    _manualDirty = false;
-
-    /* Restore section from URL hash (default: connection); replace state so the
-       initial load does not leave a spurious Back entry. */
-    var initialSection = _sectionFromHash();
-    _applySection(initialSection);
-    history.replaceState(null, '', '#' + initialSection);
-});
-
-/* ── Module Management ── */
-function initModuleToggles() {
-    document.querySelectorAll('.module-toggle-input').forEach(function(toggle) {
-        toggle.addEventListener('change', function() {
-            var toggleEl = this;
-            if (toggleEl.getAttribute('data-is-threshold') === 'true' && toggleEl.checked) {
-                document.querySelectorAll('.module-toggle-input[data-is-threshold="true"]').forEach(function(other) {
-                    if (other !== toggleEl) other.checked = false;
-                });
-            }
-            _saveSettingsInstantly();
-        });
-    });
-}
-
-document.addEventListener('DOMContentLoaded', initModuleToggles);
-
-/* ── Theme System ── */
-var _previewingThemeId = null;
-var _originalStyles = {};
-
-function previewTheme(card) {
-    var themeId = card.getAttribute('data-theme-id');
-    var isDark = document.documentElement.getAttribute('data-theme') !== 'light';
-    var mode = isDark ? 'dark' : 'light';
-    var vars = JSON.parse(card.getAttribute('data-theme-' + mode));
-
-    _originalStyles = {};
-    Object.keys(vars).forEach(function(key) {
-        _originalStyles[key] = document.documentElement.style.getPropertyValue(key);
-        document.documentElement.style.setProperty(key, vars[key]);
-    });
-
-    _previewingThemeId = themeId;
-    var overlay = document.getElementById('theme-preview-overlay');
-    var nameEl = document.getElementById('preview-theme-name');
-    if (nameEl) nameEl.textContent = card.getAttribute('data-theme-name') || themeId;
-    if (overlay) overlay.style.display = '';
-}
-
-function cancelPreview() {
-    Object.keys(_originalStyles).forEach(function(key) {
-        if (_originalStyles[key]) {
-            document.documentElement.style.setProperty(key, _originalStyles[key]);
-        } else {
-            document.documentElement.style.removeProperty(key);
-        }
-    });
-    _originalStyles = {};
-    _previewingThemeId = null;
-    var overlay = document.getElementById('theme-preview-overlay');
-    if (overlay) overlay.style.display = 'none';
-}
-
-function applyPreviewedTheme() {
-    if (_previewingThemeId) {
-        applyTheme(_previewingThemeId);
-    }
-}
-
-function applyTheme(themeId) {
-    _guardUnsaved().then(function(proceed) {
-        if (!proceed) return;
-        fetch(docsightUrl('/api/modules/' + themeId + '/enable'), { method: 'POST' })
-            .then(function(r) { return r.json(); })
-            .then(function(data) {
-                if (data.success) {
-                    localStorage.setItem('docsight-active-theme', themeId);
-                    var overlay = document.getElementById('theme-preview-overlay');
-                    if (overlay) overlay.style.display = 'none';
-                    _previewingThemeId = null;
-                    _originalStyles = {};
-                    location.reload();
-                } else {
-                    showToast(data.error || (T.theme_apply_failed || 'Failed to apply theme'), false);
-                }
-            })
-            .catch(function(err) {
-                showToast((T.error_prefix || 'Error') + ': ' + err.message, false);
-            });
-    });
-}
-
-
-/* ── Theme Registry ── */
-var _themeRegistryFetching = false;
-var _themeRegistryLoaded = false;
-
-function loadThemeRegistryIfNeeded() {
-    if (_themeRegistryLoaded || _themeRegistryFetching) return;
-    refreshRegistry();
-}
-
-function refreshRegistry() {
-    var gallery = document.getElementById('registry-gallery');
-    if (!gallery || _themeRegistryFetching) return;
-    _themeRegistryFetching = true;
-    gallery.textContent = '';
-    var p = document.createElement('p');
-    p.textContent = T.loading || 'Loading...';
-    var loading = document.createElement('div');
-    loading.className = 'empty-state';
-    loading.appendChild(p);
-    gallery.appendChild(loading);
-
-    fetch(docsightUrl('/api/themes/registry'))
-        .then(function(r) { return r.json(); })
-        .then(function(themes) {
-            gallery.textContent = '';
-            _themeRegistryLoaded = true;
-            if (!themes.length) {
-                var empty = document.createElement('div');
-                empty.className = 'empty-state';
-                var msg = document.createElement('p');
-                msg.textContent = T.themes_all_installed || 'All available themes are installed';
-                empty.appendChild(msg);
-                gallery.appendChild(empty);
-                return;
-            }
-            themes.forEach(function(theme) {
-                var card = document.createElement('div');
-                card.className = 'theme-card glass';
-
-                var info = document.createElement('div');
-                info.className = 'theme-info';
-                var name = document.createElement('div');
-                name.className = 'theme-name';
-                name.textContent = theme.name;
-                var desc = document.createElement('div');
-                desc.className = 'theme-desc';
-                desc.textContent = theme.description || '';
-                var meta = document.createElement('div');
-                meta.className = 'theme-meta';
-                var ver = document.createElement('span');
-                ver.textContent = 'v' + theme.version;
-                var auth = document.createElement('span');
-                auth.textContent = theme.author || '';
-                meta.appendChild(ver);
-                meta.appendChild(auth);
-                info.appendChild(name);
-                info.appendChild(desc);
-                info.appendChild(meta);
-
-                var actions = document.createElement('div');
-                actions.className = 'theme-actions';
-                var btn = document.createElement('button');
-                btn.type = 'button';
-                btn.className = 'btn btn-primary btn-sm';
-                btn.textContent = T.theme_install || 'Install';
-                btn.addEventListener('click', function() {
-                    installTheme(theme.id, theme.download_url);
-                });
-                actions.appendChild(btn);
-
-                card.appendChild(info);
-                card.appendChild(actions);
-                gallery.appendChild(card);
-            });
-            lucide.createIcons();
-        })
-        .catch(function() {
-            gallery.textContent = '';
-            _themeRegistryLoaded = false;
-            var err = document.createElement('div');
-            err.className = 'empty-state';
-            var msg = document.createElement('p');
-            msg.textContent = T.theme_registry_failed || 'Failed to load registry';
-            err.appendChild(msg);
-            gallery.appendChild(err);
-        })
-        .finally(function() { _themeRegistryFetching = false; });
-}
-
-function installTheme(themeId, downloadUrl) {
-    fetch(docsightUrl('/api/themes/install'), {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ id: themeId, download_url: downloadUrl }),
-    })
-        .then(function(r) { return r.json(); })
-        .then(function(data) {
-            if (data.success) {
-                showToast(T.theme_installed || 'Theme installed — restart required', true);
-                refreshRegistry();
-            } else {
-                showToast(data.error || (T.theme_install_failed || 'Install failed'), false);
-            }
-        })
-        .catch(function(err) {
-            showToast((T.error_prefix || 'Error') + ': ' + err.message, false);
-        });
-}
-
-/* ── Smart Capture History ── */
-// Note: uses innerHTML with escapeHtml() for all dynamic content, consistent with
-// the existing pattern used throughout events.js, speedtest.js, and correlation.js.
-function loadSmartCaptureHistory() {
-    var loading = document.getElementById('sc-history-loading');
-    var empty = document.getElementById('sc-history-empty');
-    var tableWrap = document.getElementById('sc-history-table-wrap');
-    var tbody = document.getElementById('sc-history-tbody');
-    if (!tbody) return;
-
-    if (loading) loading.style.display = '';
-    if (empty) empty.style.display = 'none';
-    if (tableWrap) tableWrap.style.display = 'none';
-
-    fetch(docsightUrl('/api/smart-capture/executions?limit=50'))
-        .then(function(r) { return r.json(); })
-        .then(function(data) {
-            if (loading) loading.style.display = 'none';
-            var execs = data.executions || [];
-            while (tbody.firstChild) tbody.removeChild(tbody.firstChild);
-            if (execs.length === 0) {
-                if (empty) empty.style.display = '';
-                return;
-            }
-            if (tableWrap) tableWrap.style.display = '';
-            var statusLabels = {
-                completed: T.sc_status_completed || 'Completed',
-                fired: T.sc_status_fired || 'Fired',
-                pending: T.sc_status_pending || 'Pending',
-                suppressed: T.sc_status_suppressed || 'Suppressed',
-                expired: T.sc_status_expired || 'Expired'
-            };
-            execs.forEach(function(ex) {
-                var tr = document.createElement('tr');
-                var ts = ex.created_at ? ex.created_at.replace('T', ' ').replace('Z', '') : '';
-                var trigger = escapeHtml(ex.trigger_type || '');
-                var label = statusLabels[ex.status] || ex.status;
-                var detail = '';
-                if (ex.suppression_reason) {
-                    detail = escapeHtml(ex.suppression_reason);
-                } else if (ex.linked_result_id) {
-                    detail = 'Result #' + ex.linked_result_id;
-                } else if (ex.last_error) {
-                    detail = escapeHtml(ex.last_error);
-                }
-                // Dynamic content sanitized via escapeHtml before insertion
-                tr.innerHTML = '<td style="white-space:nowrap;font-size:0.85em;">' + escapeHtml(ts) + '</td>'
-                    + '<td>' + trigger + '</td>'
-                    + '<td><span class="sc-status-' + escapeHtml(ex.status) + '">' + escapeHtml(label) + '</span></td>'
-                    + '<td style="font-size:0.85em;color:var(--muted);">' + detail + '</td>';
-                tbody.appendChild(tr);
-            });
-            if (typeof lucide !== 'undefined') lucide.createIcons();
-        })
-        .catch(function() {
-            if (loading) loading.style.display = 'none';
-            if (empty) empty.style.display = '';
-            showToast(T.sc_history_error || 'Failed to load execution history', false);
-        });
-}
-
-/* ── Smart Capture Guardrails Summary ── */
-function updateGuardrailsSummary() {
-    var el = document.getElementById('sc-guardrails-summary');
-    var cooldownEl = document.getElementById('sc_global_cooldown');
-    var maxEl = document.getElementById('sc_max_actions_per_hour');
-    var speedtestIntervalEl = document.getElementById('sc_speedtest_min_interval');
-    var speedtestDailyEl = document.getElementById('sc_speedtest_max_actions_per_day');
-    if (!el || !cooldownEl || !maxEl) return;
-    var cooldown = parseInt(cooldownEl.value) || 0;
-    var maxPerHour = parseInt(maxEl.value) || 1;
-    var speedtestInterval = speedtestIntervalEl ? (parseInt(speedtestIntervalEl.value) || 0) : 0;
-    var speedtestDaily = speedtestDailyEl ? (parseInt(speedtestDailyEl.value) || 0) : 0;
-    var cooldownStr;
-    if (cooldown >= 3600) cooldownStr = Math.round(cooldown / 3600) + 'h';
-    else if (cooldown >= 60) cooldownStr = Math.round(cooldown / 60) + ' min';
-    else cooldownStr = cooldown + 's';
-    var speedtestIntervalStr;
-    if (speedtestInterval >= 3600) speedtestIntervalStr = Math.round(speedtestInterval / 3600) + 'h';
-    else if (speedtestInterval >= 60) speedtestIntervalStr = Math.round(speedtestInterval / 60) + ' min';
-    else speedtestIntervalStr = speedtestInterval + 's';
-    var tpl = T.sc_guardrails_summary || 'At most %max%/hour, minimum %cooldown% apart; Smart Capture speedtests: %speedtestMax%/day, %speedtestInterval% apart';
-    el.textContent = tpl
-        .replace('%max%', maxPerHour)
-        .replace('%cooldown%', cooldownStr)
-        .replace('%speedtestMax%', speedtestDaily)
-        .replace('%speedtestInterval%', speedtestIntervalStr);
-}
-
-// Initialize guardrails summary on load and input
-(function() {
-    function init() {
-        updateGuardrailsSummary();
-        ['sc_global_cooldown', 'sc_max_actions_per_hour', 'sc_speedtest_min_interval', 'sc_speedtest_max_actions_per_day'].forEach(function(id) {
-            var el = document.getElementById(id);
-            if (el) el.addEventListener('input', updateGuardrailsSummary);
-        });
-    }
-    if (document.readyState === 'loading') {
-        document.addEventListener('DOMContentLoaded', init);
-    } else {
-        init();
-    }
-})();
-
-/* ── Community Module Registry ── */
-var _registryFetching = false;
-
-function refreshModuleRegistry() {
-    var gallery = document.getElementById('module-registry-gallery');
-    var empty = document.getElementById('module-registry-empty');
-    var loading = document.getElementById('module-registry-loading');
-    if (!gallery || _registryFetching) return;
-    _registryFetching = true;
-
-    if (loading) loading.style.display = '';
-    if (empty) empty.style.display = 'none';
-    gallery.style.display = 'none';
-
-    fetch(docsightUrl('/api/modules/registry'))
-        .then(function(r) { return r.json(); })
-        .then(function(modules) {
-            if (loading) loading.style.display = 'none';
-            if (!modules || modules.length === 0) {
-                gallery.style.display = 'none';
-                if (empty) empty.style.display = '';
-                return;
-            }
-            gallery.style.display = '';
-            if (empty) empty.style.display = 'none';
-
-            // Build cards — dynamic content assigned via textContent (inherently safe)
-            _renderRegistryCards(gallery, modules);
-            if (typeof lucide !== 'undefined') lucide.createIcons();
-        })
-        .catch(function() {
-            if (loading) loading.style.display = 'none';
-            gallery.style.display = 'none';
-            if (empty) empty.style.display = '';
-            showToast(T.extensions_fetch_failed || 'Failed to load registry.', false);
-        })
-        .finally(function() { _registryFetching = false; });
-}
-
-function _renderRegistryCards(gallery, modules) {
-    while (gallery.firstChild) gallery.removeChild(gallery.firstChild);
-
-    modules.forEach(function(mod) {
-        var status = mod.status || 'not_installed';
-
-        var card = document.createElement('div');
-        card.className = 'module-registry-card';
-
-        var info = document.createElement('div');
-        info.className = 'registry-card-info';
-
-        var nameRow = document.createElement('div');
-        nameRow.className = 'registry-card-name';
-        nameRow.textContent = mod.name || mod.id;
-        if (mod.verified) {
-            var vBadge = document.createElement('span');
-            vBadge.className = 'module-badge badge-verified';
-            vBadge.textContent = T.extensions_verified || 'Verified';
-            nameRow.appendChild(vBadge);
-        }
-
-        var desc = document.createElement('div');
-        desc.className = 'registry-card-desc';
-        desc.textContent = mod.description || '';
-
-        var meta = document.createElement('div');
-        meta.className = 'registry-card-meta';
-        meta.textContent = (mod.author || '') + ' \u00B7 v' + (mod.version || '') + ' \u00B7 ';
-        var statusSpan = document.createElement('span');
-        if (status === 'not_installed') {
-            statusSpan.textContent = T.extensions_not_installed || 'Not installed';
-        } else if (status === 'installed_disabled') {
-            statusSpan.textContent = T.extensions_installed || 'Installed';
-            statusSpan.className = 'registry-status-warn';
-        } else {
-            statusSpan.textContent = T.extensions_installed || 'Installed';
-            statusSpan.className = 'registry-status-good';
-        }
-        meta.appendChild(statusSpan);
-
-        info.appendChild(nameRow);
-        info.appendChild(desc);
-        info.appendChild(meta);
-
-        var action = document.createElement('div');
-        action.className = 'registry-card-action';
-
-        var btn = document.createElement('button');
-        if (status === 'not_installed') {
-            btn.className = 'btn btn-sm btn-install';
-            btn.textContent = T.extensions_install || 'Install';
-            btn.addEventListener('click', function(e) { installModule(e, mod.id, mod.download_url); });
-        } else {
-            btn.className = 'btn btn-sm btn-uninstall';
-            btn.textContent = T.extensions_uninstall || 'Uninstall';
-            btn.addEventListener('click', function(e) { uninstallModule(e, mod.id); });
-        }
-        action.appendChild(btn);
-
-        card.appendChild(info);
-        card.appendChild(action);
-        gallery.appendChild(card);
-    });
-}
-
-function _runModuleAction(e, id, action, downloadUrl) {
-    var btn = e.currentTarget;
-    var installing = action === 'install';
-    var idleClass = installing ? 'btn-install' : 'btn-uninstall';
-    var idleText = installing
-        ? (T.extensions_install || 'Install')
-        : (T.extensions_uninstall || 'Uninstall');
-    var failedText = installing
-        ? (T.extensions_install_failed || 'Installation failed')
-        : (T.extensions_uninstall_failed || 'Uninstall failed');
-    if (btn.dataset.confirmPending !== 'true') {
-        btn.dataset.confirmPending = 'true';
-        btn.className = 'btn btn-sm btn-confirm';
-        btn.textContent = T.extensions_confirm || 'Confirm?';
-        btn._resetTimer = setTimeout(function() {
-            btn.dataset.confirmPending = 'false';
-            btn.className = 'btn btn-sm ' + idleClass;
-            btn.textContent = idleText;
-        }, 3000);
-        return;
-    }
-    clearTimeout(btn._resetTimer);
-    btn.dataset.confirmPending = 'false';
-    btn.disabled = true;
-    btn.textContent = '...';
-
-    var payload = {id: id};
-    if (installing) payload.download_url = downloadUrl;
-    fetch(docsightUrl('/api/modules/' + action), {
-        method: 'POST',
-        headers: {'Content-Type': 'application/json'},
-        body: JSON.stringify(payload),
-    })
-    .then(function(r) { return r.json(); })
-    .then(function(data) {
-        if (data.success) {
-            showToast(
-                installing
-                    ? (T.extensions_install_success || 'Module installed successfully')
-                    : (T.extensions_uninstall_success || 'Module uninstalled'),
-                true
-            );
-            var banner = document.getElementById('module-restart-banner');
-            if (banner) { banner.style.display = ''; if (typeof lucide !== 'undefined') lucide.createIcons({nodes: [banner]}); }
-            refreshModuleRegistry();
-        } else {
-            showToast(data.error || failedText, false);
-            btn.disabled = false;
-            btn.className = 'btn btn-sm ' + idleClass;
-            btn.textContent = idleText;
-        }
-    })
-    .catch(function(err) {
-        showToast(failedText + ': ' + err.message, false);
-        btn.disabled = false;
-        btn.className = 'btn btn-sm ' + idleClass;
-        btn.textContent = idleText;
-    });
-}
-
-/* Thin public entry points retained for existing card click handlers. */
-function installModule(e, id, downloadUrl) {
-    _runModuleAction(e, id, 'install', downloadUrl);
-}
-
-function uninstallModule(e, id) {
-    _runModuleAction(e, id, 'uninstall');
-}
+})(DOCSightSettings);
