+++ tests/test_static_contracts.py
@@ -409,7 +409,7 @@ def test_service_worker_precache_references_existing_public_assets() -> None:
 def test_dead_static_js_helpers_stay_removed() -> None:
     assert not (STATIC / "js" / "icons.js").exists()
 
-    settings_js = (STATIC / "js" / "settings.js").read_text(encoding="utf-8")
+    settings_js = "\n".join(path.read_text(encoding="utf-8") for path in (STATIC / "js" / "settings").glob("*.js"))
     utils_js = (STATIC / "js" / "utils.js").read_text(encoding="utf-8")
     sw_js = (STATIC / "sw.js").read_text(encoding="utf-8")
     templates = "\n".join(
@@ -560,7 +560,7 @@ def test_modal_consumers_have_no_absent_api_fallbacks() -> None:
 
 
 def test_frontend_simplifications_keep_single_owners() -> None:
-    settings = (STATIC / "js" / "settings.js").read_text(encoding="utf-8")
+    settings = (STATIC / "js" / "settings" / "module-registry.js").read_text(encoding="utf-8")
     connection_charts = (
         MODULES
         / "connection_monitor"
